Major changes in log store truncation logic (#572)
* Major changes in log store truncation logic

Here are the several changes done in log store truncation and super block

* The present logstore had serious issue of acting upon batch completion of logdev flush
only for the last logstore. As a result, there are chances that truncation lags way
behind and also results in error in trunction. This is fixed by maintaining all logstores
participated in the logdev flush and notify them about batch_completion for each log store

* Log device truncation fails to move forward in case a particular logstore didn't have
any writes. This is because we take lowest of all and since a logstore doesn't move forward
if there are no writes, the truncation never moves forward and soon logdev runs out of space.
This is fixed by doing 2 phase truncation, in first phase get to know all about truncation
and also keep track of if there are any pending writes. If none, then allow that logstore
not to be part of the current truncation.

* Log device metadata is stored in 2 places and worse some data is written in one and read
in other. As a result, the logdev replay always starts with 0. Made logdev use metablk and
single source of truth and removed the persistence in virtual dev

* As a result of above fixes, lots of places had some boundary condition failures, especially
in clean restart cases, fixed all of them.

* Log device is made self sustained in terms of its truncation point. Logdev has its own
metadata (meta blk) and it will contain each logstore meta which stores its own truncation point.
This is persisted everytime we do device truncation. As a result, it does not depend on the caller
to provide lsns, start seq, truncate etc and fixed several test case failures.

* Logdev stop now waits for flush or truncate to be completed and do not allow flush while stopping

* Enabled the logstore restartability test cases